+++ nf_core/modules/lint/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Code for linting modules in the nf-core/modules repository and
 in nf-core pipelines
@@ -21,10 +20,9 @@ from rich.table import Table
 
 import nf_core.modules.modules_utils
 import nf_core.utils
+from nf_core.components.components_command import ComponentCommand
 from nf_core.lint_utils import console
-from nf_core.modules.modules_command import ModuleCommand
 from nf_core.modules.modules_json import ModulesJson
-from nf_core.modules.modules_repo import ModulesRepo
 from nf_core.modules.nfcore_module import NFCoreModule
 from nf_core.utils import plural_s as _s
 
@@ -48,7 +46,7 @@ class LintResult(object):
         self.module_name = mod.module_name
 
 
-class ModuleLint(ModuleCommand):
+class ModuleLint(ComponentCommand):
     """
     An object for linting modules either in a clone of the 'nf-core/modules'
     repository or in any nf-core pipeline directory
@@ -73,7 +71,9 @@ class ModuleLint(ModuleCommand):
         no_pull=False,
         hide_progress=False,
     ):
-        super().__init__(dir=dir, remote_url=remote_url, branch=branch, no_pull=no_pull, hide_progress=False)
+        super().__init__(
+            "modules", dir=dir, remote_url=remote_url, branch=branch, no_pull=no_pull, hide_progress=hide_progress
+        )
 
         self.fail_warned = fail_warned
         self.passed = []