+++ nf_core/utils.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Common utility functions for the nf-core python package.
 """